==> notebooks/03_Train_Data.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import r2_score
import pandas as pd

df2020 = pd.read_csv("../data/3/spotify_2020_cleaned.csv")
# df2023 = pd.read_csv("../data/3/spotify_2023_cleaned.csv")


x2020 = df2020.drop(columns=['id', 'name', 'release_date', 'artists', 'id_artists', 'popularity', 'explicit', 'artist_id', 'merge_key'])
# x2023 = df2023.drop(columns=['id', 'name', 'release_date', 'popularity'])

y2020 = df2020['popularity']
# y2023 = df2023['popularity']

X_train, X_test, y_train, y_test = train_test_split(x2020, y2020, test_size=0.2, random_state=22)
model = RandomForestRegressor()
model.fit(X_train, y_train)
preds = model.predict(X_test)
print("R²:", r2_score(y_test, preds))
